% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsParser.R
\name{argsParser}
\alias{argsParser}
\title{Parser of an argument string}
\usage{
argsParser(option, args, sep = " ", novalue_response = NULL)
}
\arguments{
\item{option}{character strings containg options (or flag) whose values}

\item{args}{String containing all the arguments of an R script}

\item{sep}{separator character. Default is \code{" "}. If it is of length 2, the first is seperator among different options, the second is betwwen option name and its value.}

\item{novalue_response}{value used in case the option is missing. Default is \code{NULL}.}
}
\description{
This command parses ...DESCRITION TO DO !!!
}
\examples{

args <- "--value  6  --fruit apple"

option <- "--fruit"

value <- argsParser(option=option,args=args)

option2 <- "--jobs"

value2 <- argsParser(option=option2,args=args)
value22 <- argsParser(option=option2,args=args,novalue_response="./")
 args_b <- "value=6 , fruit=apple"
option3 <- "value"
value <- argsParser(option=option3,args=args_b,sep=c(",","="))

}

