\name{color.bar.raster}
\alias{color.bar.raster}
\title{Graphic Representation of a Color legend of a Raster or GeotopbrickRaster object as a Color bar, inspired by the function written by John Colby}
\usage{
color.bar.raster(x, col, ...)
}
\arguments{
  \item{x}{a Rster or GeotopRasterBrick object}

  \item{col}{the color palette used}

  \item{...}{arguments to be passed to
  \code{\link{color.bar}}}
}
\description{
Graphic Representation of a Color legend of a Raster or
GeotopbrickRaster object as a Color bar, inspired by the
function written by John Colby
}
\seealso{
\code{\link{color.bar}}
}

