% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{geotopbricks-package}
\alias{geotobricks}
\alias{geotopbricks-package}
\title{geotopbricks: Analyzes raster maps as input/output files from the Hydrological Distributed Model GEOtop}
\description{
This packages uses R raster utilities to read and analize outputs of the Distributed Hydrological Model GEOtop \url{www.geotop.org}.
It contains functions and methods to import maps and other keywords from geotop.inpts file.
Any information about the GEOtop Distributed Hydrological Model is available on www.geotop.org.
Two examples are shown: \url{http://meteogis.fmach.it/idroclima/panola13_run2xC_test3/} and \url{http://meteogis.fmach.it/idroclima/ton-toss/}.
These examples are tested on two simulation cases run with GEOtop built 1.225-9 mostly developed by Stefano Endrizzi (\url{http://www.geo.uzh.ch/en/units/physical-geography-3g/about-us/staff/stefano-endrizzi}).
Bugs/comments/questions/collaboration of any kind are warmly welcomed.
}
\details{
\tabular{ll}{
Package: \tab geotopbricks\cr
Type: \tab Package\cr
Version: \tab 1.3.5\cr
Date: \tab   2013-08-25\cr
License: \tab GPL (>= 2)

\cr
LazyLoad: \tab yes\cr
Depends:  zoo,rgdal,methods,stringr,raster,soilwater \cr
}
}
\note{
geotobricks is an on-going project. All criticism, comments and suggestions are well welcomed.

geotobricks is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
   (at your option) any later version.

   geotopbricks is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.

   You should have received a copy of the GNU General Public License
   along with this program.  If not, see \url{http://www.gnu.org/licenses/}.
}
\author{
Emanuele Cordano \email{emanuele.cordano@gmail.com}, Daniele Andreis, Fabio Zottele.
}
\references{
\url{wwww.geotop.org}
}
\keyword{GIS,}
\keyword{hydrology}
\keyword{package,}

