% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_value.R
\name{max_value}
\alias{max_value}
\title{max_value}
\usage{
max_value(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{GeotopRasterBrick}} object.}

\item{na.rm, ...}{further arguments for \code{\link{max}}.}
}
\value{
the maximum (scalar) values of a \code{\link{GeotopRasterBrick}} object
}
\description{
Gets the maximum (scalar) values of a \code{\link{GeotopRasterBrick}} object
}
