% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_value.R
\name{min_value}
\alias{min_value}
\title{min_value}
\usage{
min_value(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{GeotopRasterBrick}} object.}

\item{na.rm, ...}{further arguments foe \code{\link{min}}.}
}
\value{
the minimum (scalar) values of a \code{\link{GeotopRasterBrick}} object
}
\description{
Gets the minimum (scalar) values of a \code{\link{GeotopRasterBrick}} object
}
