\docType{methods}
\name{Ops}
\alias{Ops}
\alias{Ops,GeotopRasterBrick-method}
\title{Ops}
\arguments{
  \item{e1,e2}{the \code{\link{GeotopRasterBrick}} or
  numeric objects}
}
\description{
  \code{Ops} method for a \code{GeotopRasterBrick} object
}
\note{
  If \code{e1} or \code{e2} time index is not taken into
  account.
}
\keyword{methods}

