% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.ascii.vectorized.brick.R
\name{read.ascii.vectorized.brick}
\alias{read.ascii.vectorized.brick}
\title{Read a text file containing  values and matedata of a z-layer brick referred to a time instant (e.g. date). The file is formatted like an ascii format like \code{'geotop.inpts'} file.}
\usage{
read.ascii.vectorized.brick(file = NULL, comment = "!", crs = "",
  NAflag = -9999, matlab.syntax = FALSE, ...)
}
\arguments{
\item{file}{file name to write}

\item{comment}{character. Comment indicator. Default is \code{"!"}.}

\item{crs}{Character or object of class CRS. PROJ4 type description of a Coordinate Reference System (map projection) (optional). See \code{\link{brick}} or \code{\link{raster}}.}

\item{NAflag}{numeric. Dafauli is -9999, see \code{\link{writeRasterxGEOtop}}.}

\item{matlab.syntax}{logical value. Default is \code{FALSE}. If \code{TRUE} the file syntax is like the one of a *.m Matlab script file.}

\item{...}{further aguments inserted as attribute}
}
\value{
the \code{\link{RasterBrick-class}} object
}
\description{
Read a text file containing  values and matedata of a z-layer brick referred to a time instant (e.g. date). The file is formatted like an ascii format like \code{'geotop.inpts'} file.
}
\examples{
# see the examples of read.ascii.vectorized.brick
}
\seealso{
\code{\link{write.ascii.vectorized.brick}}
}

