% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.recovery.state.R
\name{set.geotop.recovery.state}
\alias{set.geotop.recovery.state}
\title{This function re-writes the recovery ascii raster maps in a given folder}
\usage{
set.geotop.recovery.state(rec, newRecFolder, ...)
}
\arguments{
\item{rec}{a \code{list} object returened by \code{\link{get.geotop.recovery.state}}}

\item{newRecFolder}{directory where to write all recovery raster asccii maps}

\item{...}{further arguments}
}
\description{
This function re-writes the recovery ascii raster maps in a given folder
}
\examples{
# See the examples of the 'get.geotop.recovery.state' function
}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{get.geotop.recovery.state}},\code{\link{writeRasterxGEOtop}}
}

