% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\name{zoo-class}
\alias{zoo-class}
\title{A GeotopRasterBrick: an object to manage raster maps provied by GEOtop!!}
\description{
A GeotopRasterBrick: an object to manage raster maps provied by GEOtop!!
}
\examples{
showClass("zoo")
}
