% Generated by roxygen2 (4.0.2): do not edit by hand
\name{declared.geotop.inpts.keywords}
\alias{declared.geotop.inpts.keywords}
\title{Collects all keywords contained in the 'getop.inpts' configuration files and their values in a data frame object.}
\usage{
declared.geotop.inpts.keywords(wpath, inpts.file = "geotop.inpts",
  comment = "!", exceptions = "Date", warn = FALSE,
  no.comment = c("!>!", "!>>!"), ...)
}
\arguments{
\item{wpath}{working directory containing GEOtop files}

\item{inpts.file}{name of the GEOtop configuration file. Default is \code{"geotop.inpts"}}

\item{comment}{comment indicator charcater. Default is \code{"!"}}

\item{exceptions}{string vector. If keywords contain an element of this vector, the blank spaces in Value \code{" "} will not be removed.}

\item{warn}{logical argument of \code{\link{readLines}}. Default is \code{FALSE}.}

\item{no.comment}{string indicatos read as comment ones by GEOtop but they do not indicate comments by "geotopbricks" package.}

\item{...}{further arguments of \code{\link{readLines}}}
}
\value{
a data frame with two columns: \code{Keyword} and \code{Value}
}
\description{
Collects all keywords contained in the 'getop.inpts' configuration files and their values in a data frame object.
}
\seealso{
\code{\link{get.geotop.inpts.keyword.value}}
}

