\name{max_value}
\alias{max_value}
\title{max_value}
\usage{
max_value(x)
}
\arguments{
  \item{x}{a \code{\link{GeotopRasterBrick}} object}

  \item{...}{further arguments}
}
\value{
the maximum (scalar) values of a
\code{\link{GeotopRasterBrick}} object
}
\description{
Gets the maximum (scalar) values of a
\code{\link{GeotopRasterBrick}} object
}

