% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.GeotopRasterBrick.R
\docType{methods}
\name{Ops}
\alias{Ops}
\alias{Ops,GeotopRasterBrick,GeotopRasterBrick-method}
\alias{Ops,GeotopRasterBrick,numeric-method}
\alias{Ops,numeric,GeotopRasterBrick-method}
\title{Ops}
\usage{
\S4method{Ops}{GeotopRasterBrick,GeotopRasterBrick}(e1, e2)

\S4method{Ops}{GeotopRasterBrick,numeric}(e1, e2)

\S4method{Ops}{numeric,GeotopRasterBrick}(e1, e2)
}
\arguments{
\item{e1,e2}{the \code{\link{GeotopRasterBrick}} or numeric  objects}
}
\description{
\code{Ops} method for a \code{GeotopRasterBrick} object




}
\note{
If \code{e1} or \code{e2} time index is not taken into account.
}
\keyword{methods}

