% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zoo-class}
\alias{zoo-class}
\title{A GeotopRasterBrick: an object to manage raster maps provied by GEOtop!!}
\description{
A GeotopRasterBrick: an object to manage raster maps provied by GEOtop!!
}
\examples{
showClass("zoo")
}
\keyword{classes}

